+++ memtable/vectorrep.cc
@@ -92,7 +92,8 @@ class VectorRep : public MemTableRep {
   };
 
   // Return an iterator over the keys in this representation.
-  MemTableRep::Iterator* GetIterator(Arena* arena) override;
+  MemTableRep::Iterator* GetIterator(Arena* arena,
+                                     bool part_of_flush = false) override;
 
  private:
   friend class Iterator;
@@ -263,7 +264,8 @@ void VectorRep::Get(const LookupKey& k, void* callback_args,
   }
 }
 
-MemTableRep::Iterator* VectorRep::GetIterator(Arena* arena) {
+MemTableRep::Iterator* VectorRep::GetIterator(Arena* arena,
+                                              bool /*part_of_flush*/) {
   char* mem = nullptr;
   if (arena != nullptr) {
     mem = arena->AllocateAligned(sizeof(Iterator));